% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cor.R
\name{calc_cor}
\alias{calc_cor}
\title{Calculate correlation between two vectors}
\usage{
calc_cor(x_data, y_data, method, legend_pos, legend_col = "blue")
}
\arguments{
\item{legend_pos}{}
}
\description{
Calculate correlation between two vectors
}
