% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{contour_legend}
\alias{contour_legend}
\title{Do a filled contour legend}
\usage{
contour_legend(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out
  = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y, finite =
  TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels),
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, ...)
}
\arguments{
\item{...}{}
}
\description{
Do a filled contour legend
}
