% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_abline.R
\name{draw.abline}
\alias{draw.abline}
\title{Function for drawing a linear regression through data}
\usage{
draw.abline(x_data, y_data, interval = "c", col = "blue",
  legend_pos = "topright", legend_col = col, ...)
}
\arguments{
\item{...}{}
}
\description{
Function for drawing a linear regression through data
}
