% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_loess.R
\name{draw.loess}
\alias{draw.loess}
\title{Function for drawing a loess fit through data}
\usage{
draw.loess(x_data, y_data, model_sp = 0, model_bw = 0, model_dg,
  bs_repeats = 1000, fit_resolution = 100, conf = 0.68, col = "blue",
  lwd = 2, ...)
}
\arguments{
\item{col}{}
}
\description{
Function for drawing a loess fit through data
}
