% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_abline.R
\name{draw.princomp}
\alias{draw.princomp}
\title{Function for drawing a linear regression through data}
\usage{
draw.princomp(x_data, y_data, col = "blue", interval = "c",
  legend_pos = "topright", legend_col = "blue", x_fit_data = seq(from =
  min(x_data), to = max(x_data), length.out = 200), ...)
}
\arguments{
\item{...}{}
}
\description{
Function for drawing a linear regression through data
}
