% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{fade.cols}
\alias{fade.cols}
\title{Function for fading a colour by a given alpha}
\usage{
fade.cols(col, alpha)
}
\arguments{
\item{alpha}{}
}
\description{
Function for fading a colour by a given alpha
}
