% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_abline.R
\name{get_prcomp_abline}
\alias{get_prcomp_abline}
\title{Fit princomp line}
\usage{
get_prcomp_abline(x, y)
}
\arguments{
\item{x}{The x data}

\item{y}{The y data}
}
\value{
Returns the intercept and slope.
}
\description{
Get the intercept and slope for the 1st principle component.
}
