% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_stacked.R
\name{hist_stacked}
\alias{hist_stacked}
\title{Plot stacked histogram}
\usage{
hist_stacked(data, breaks = "Sturges", ylim = NULL, xlab = "",
  main = "", hist_cols = rainbow(length(data)), border = "black",
  legend_names, legend_inset = 0.05)
}
\arguments{
\item{legend_inset}{}
}
\description{
Plot stacked histogram
}
