% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{is.multiple}
\alias{is.multiple}
\title{Check if x is a multiple of something}
\usage{
is.multiple(x, multiple)
}
\arguments{
\item{multiple}{}
}
\description{
Check if x is a multiple of something
}
