% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{labbook.init}
\alias{labbook.init}
\title{Initiate a new labbook}
\usage{
labbook.init(
  labbook.dir,
  labbook.title,
  project.title,
  labbook.dirname = "labbook"
)
}
\arguments{
\item{labbook.dir}{The directory in which to create the labbook (defaults to the current working directory)}

\item{labbook.title}{The name for the labbook (e.g. "Sam's labbook")}

\item{project.title}{The name for the first project}

\item{labbook.dirname}{The directory name for your labbook (default: "labbook")}
}
\description{
Initiate a new labbook
}
