% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{labbook.newPage}
\alias{labbook.newPage}
\title{Create a new labbook page}
\usage{
labbook.newPage(
  filename,
  project.dir = NULL,
  openfile = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{project.dir}{}
}
\description{
Create a new labbook page
}
