% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_labBook_files.R
\name{labBook_removePage}
\alias{labBook_removePage}
\title{Remove labBook page}
\usage{
labBook_removePage(page_title, project_dir)
}
\arguments{
\item{project}{}
}
\description{
Removes the code file, html file and index reference for a labBook page
}
