% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labBook.R
\name{labBook_setup}
\alias{labBook_setup}
\title{Setup LabBook}
\usage{
labBook_setup(labBook_name, target_dir, project_name)
}
\arguments{
\item{target_dir}{}
}
\description{
Function to initialise a labBook
}

