% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labBook.R
\name{labBook_update}
\alias{labBook_update}
\title{Update LabBook}
\usage{
labBook_update(labBook_dir)
}
\arguments{
\item{clean_first}{}
}
\description{
Function to update lab book with new features and javascript etc.
}
