% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{layout_pairplot}
\alias{layout_pairplot}
\title{Layout a pair plot}
\usage{
layout_pairplot(x, widths = rep(1, x), heights = rep(1, x))
}
\arguments{
\item{heights}{}
}
\description{
Layout a pair plot
}
