% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{make_latex_table}
\alias{make_latex_table}
\title{Format a table as latex}
\usage{
make_latex_table(R_table)
}
\arguments{
\item{R_table}{}
}
\value{
Returns a character string of the table formatted into latex
}
\description{
Format a table as latex
}
