% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{mean_ci}
\alias{mean_ci}
\title{Function for calculating the confidence interval of a sample mean}
\usage{
mean_ci(x, conf.level = 0.95)
}
\arguments{
\item{conf.level}{}
}
\description{
Function for calculating the confidence interval of a sample mean
}
