% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mix.cols}
\alias{mix.cols}
\title{Function for mixing two colours together}
\usage{
mix.cols(col1, col2, prop = 0.5)
}
\arguments{
\item{prop}{}
}
\description{
Function for mixing two colours together
}
