% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{openTodo}
\alias{openTodo}
\title{Open the todo page}
\usage{
openTodo(openmd = TRUE, todopath = NULL)
}
\arguments{
\item{todopath}{The path to the todo file}
}
\description{
Render and open the overall todo page
}
