% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Do a 2D density function plot}
\usage{
plot_density(x, y, xlab = "", ylab = "", xaxs = "r", yaxs = "r",
  xlim = range(x), ylim = range(y), dens_col = topo.colors(40), dens_fun,
  fit_h, ndens = 100)
}
\arguments{
\item{ylab}{}
}
\description{
Do a 2D density function plot
}
