% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_text.R
\name{plot_text}
\alias{plot_text}
\title{Plot text on a blank plot}
\usage{
plot_text(text2plot, cex = 2, plot_x = 0, plot_y = 0, ...)
}
\arguments{
\item{valign}{}
}
\description{
Plot text on a blank plot
}
