% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_titres.R
\name{plot_titres}
\alias{plot_titres}
\title{Plot titres}
\usage{
plot_titres(x_data, y_data, xlim = c(-1, 8), ylim = c(-1, 8), xlab = "",
  ylab = "", nd_titres = TRUE, princomp = FALSE, princomp_col = "red",
  titre_class_means = FALSE, draw_eq_line = TRUE, lm = FALSE,
  lm_col = "blue", fit_x_nd = !nd_titres, fit_y_nd = !nd_titres,
  jitter = NA, x_jitter = jitter, y_jitter = jitter, title, ...)
}
\arguments{
\item{lm_col}{}
}
\description{
Plot titres
}
