% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{render.note}
\alias{render.note}
\title{Render a new note}
\usage{
render.note(
  notepath = NULL,
  pagepath = NULL,
  pagelink = NULL,
  project.path = NULL,
  add_index_link = TRUE,
  notetitle = NULL,
  openpage = FALSE,
  headcontent = NULL,
  headercontent = NULL
)
}
\description{
Render a new note
}
