% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{render.todo}
\alias{render.todo}
\title{Render a new todo}
\usage{
render.todo(todopath = NULL, pagepath = NULL, pagelink = NULL,
  project.path = NULL, add_index_link = TRUE,
  update_index_todo = TRUE, openpage = FALSE, headcontent = NULL)
}
\description{
Render a new todo
}
