% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf.R
\name{save2pdfNpng}
\alias{save2pdfNpng}
\title{Copies content of current device to PDF and PNG. (created by Nick)}
\usage{
save2pdfNpng(filename, widthPDF = 7, heightPDF = 7, widthPNG = 600,
  heightPNG = 400, openhtml = T, html_width = 500, overwrite = F, ...)
}
\arguments{
\item{filename}{Please supply the filename without any extensions.}

\item{...}{}
}
\description{
Function to create a pdf file of the content of the current device and also make a png copy of it, copying the link to your clipboard. Simultaneously opens the corresponding html file, provided one is currently in the project folder of the labBook.
}
\examples{
Here you can write examples.
}
