% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{setup.plot}
\alias{setup.plot}
\title{Setup a blank plot}
\usage{
setup.plot(xlim = c(-1, 1), ylim = c(-1, 1), xlab, ylab, main, margins,
  pad_axes = TRUE, x_axis = TRUE, y_axis = TRUE, xlab_dist = 2.8,
  ylab_dist = 3, y_las = 1)
}
\arguments{
\item{xlim}{The limits of the x axis.}

\item{ylim}{The limits of the y axis.}

\item{xlab}{Optional x axis label.}

\item{ylab}{Optional x axis label.}

\item{main}{Optional title.}

\item{margins}{Option to specify plot margins in inches specifically.}

\item{pad_axes}{Logical, should the plot pad x and y axes.}

\item{y_las}{Rotation of y axis tick labels.}

\item{label_axes}{Logical, should axes be labelled.}
}
\description{
This function sets up a blank plot ready for plotting.
}
