% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdf.R
\name{write.pdf}
\alias{write.pdf}
\title{Write to a pdf file}
\usage{
write.pdf(file = NA, width = 7, height = 7, html.width = 1200,
  add2cp = F, save2cp = T, ...)
}
\arguments{
\item{...}{}
}
\description{
Function to write to a pdf file and also make a png copy of it, copying the link to your clipboard.
}
