% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{labbook.newProject}
\alias{labbook.newProject}
\title{Create a new labbook project}
\usage{
labbook.newProject(project.title, project.dir = NULL, labbook.path)
}
\arguments{
\item{project.title}{The project title}

\item{project.dir}{The name of the project directory}

\item{labbook.path}{The path to the labbook}
}
\description{
Create a new labbook project
}
