% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labBook_page.R
\name{labBook_newPage}
\alias{labBook_newPage}
\title{New LabBook page}
\usage{
labBook_newPage(page.title, code.file, project, open.files = TRUE)
}
\arguments{
\item{open.files}{}
}
\description{
Function for creating new LabBook page.
}

