% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labBook_project.R
\name{labBook_newProject}
\alias{labBook_newProject}
\title{New labBook project}
\usage{
labBook_newProject(project.title)
}
\arguments{
\item{project.title}{}
}
\description{
Function to setup a new labBook project.
}

