% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labBook.R
\name{labBook_setup}
\alias{labBook_setup}
\title{Setup LabBook}
\usage{
labBook_setup(labBook_name, target_dir, project_name, open_project = TRUE)
}
\arguments{
\item{labBook_name}{Name of your LabBook}

\item{target_dir}{Directory in which to place your labBook}

\item{project_name}{Name of your first project}
}
\description{
Function to initialise a labBook
}
