% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{plot_mean_ci}
\alias{plot_mean_ci}
\title{Plot mean point with confidence intervals}
\usage{
plot_mean_ci(x, values, conf.level = 0.95, pch = 16, bar_width, min_crop,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Plot mean point with confidence intervals
}
